+++ lib/twenv/commands/archive_likes.rb
@@ -20,21 +20,23 @@ class TWEnv::ArchiveLikes < TWEnv::Command
   def process(user)
     @user = user
     @path = File.join command_storage_path, "#{user}.#{opts[:format].downcase}"
-    write_file @path, [], opts[:format]
+    write_file @path, [], 'json'
     perform_action_on_tweets method(:read_tweets),
                              method(:archive_tweet),
                              method(:print_total)
+  rescue Interrupt
+    line.end_line
+  ensure
     puts "Archive saved to #{@path}"
   end
 
   def options(slop)
-    slop.on :m, :max=, 'The maximum number of tweets to archive. Default is 50', default: 50, as: :integer
-    slop.on :f, :format=, 'The format to store the timeline in (eg json, yaml). Default is json', default: 'json', as: :string
-    slop.on :'outbound-links-only', 'Only archive tweets that link to somewhere outside Twitter', default: false, as: :boolean
-    slop.on :'no-media', "Only archive tweets that don't include media (eg video, images)", default: false
-    slop.on :'media-only', "Only archive tweets that do include media (eg video, images)", default: false
-    slop.on :'no-links', "Only archive tweets that don't include links", default: false, as: :boolean
-    slop.on :'links-only', "Only archive tweets that do include links", default: false, as: :boolean
+    slop.on :m, :max=, 'The maximum number of likes to archive. Default is unlimited', default: 0, as: :integer
+    slop.on :'outbound-links-only', 'Only archive likes that link to somewhere outside Twitter', default: false, as: :boolean
+    slop.on :'no-media', "Only archive likes that don't include media (ie: video, images)", default: false
+    slop.on :'media-only', "Only archive likes that include media (ie: video, images)", default: false
+    slop.on :'no-links', "Only archive likes that don't include links", default: false, as: :boolean
+    slop.on :'links-only', "Only archive likes that include links", default: false, as: :boolean
   end
 
   private
@@ -52,13 +54,13 @@ class TWEnv::ArchiveLikes < TWEnv::Command
 
   def print_total(total)
     line.rewind.print "#{total} likes archived"
-    throw(:cancel) if total == opts[:max]
+    throw(:cancel) if opts[:max].nonzero? && total == opts[:max]
   end
 
   def archive_tweet(tweet)
-    tweets = parse_file @path, opts[:format]
+    tweets = parse_file @path, 'json'
     tweets.push format_tweet(tweet)
-    write_file @path, tweets, opts[:format]
+    write_file @path, tweets, 'json'
   end
 
   def filter_tweets(tweets)